// frontend/src/app/shop/models/ShopOrderModel.ts
export type Order = {
    id: number
    orderName: string
    orderNumber: string
    requestMessage: string
    userId: number
    productId: number
    productCount: number
    amount: number
    basicAddress: string
    detailAddress: string
    createdAt:string
    productImg: string
    productCategory: string
    status: string
}